add authentification

+++ app/apis/tweets.py
@@ -1,7 +1,8 @@
 from flask_restplus import Namespace, Resource, fields
-from flask import abort
-from app.models import Tweet
-from app import db
+from flask import abort, request
+from app.models import Tweet, User
+from app import db, login_manager
+import base64
 
 api = Namespace('tweets')
 
@@ -38,21 +39,34 @@ class TweetResource(Resource):
     @api.marshal_with(json_tweet, code=200)
     @api.expect(json_new_tweet, validate=True)
     def patch(self, id):
+        user = load_user_from_request(request)
+        if user is None:
+            return api.abort(401, "You need to provide a valid api_key in the header or request url")
         tweet = db.session.query(Tweet).get(id)
         if tweet is None:
             api.abort(404, "Tweet {} doesn't exist".format(id))
         else:
+            if user.id != tweet.user_id:
+                api.abort(401, "You are not the owner of this tweet")
             tweet.text = api.payload["text"]
+            db.session.commit()
             return tweet
 
     def delete(self, id):
+        user = load_user_from_request(request)
+        if user is None:
+            return api.abort(401, "You need to provide a valid api_key in the header or request url")
+        text = api.payload["text"]
         tweet = db.session.query(Tweet).get(id)
         if tweet is None:
             api.abort(404, "Tweet {} doesn't exist".format(id))
         else:
+            if user.id != tweet.user_id:
+                api.abort(401, "You are not the owner of this tweet")
             db.session.delete(tweet)
             db.session.commit()
             return None
+
 
 @api.route('')
 class TweetsResource(Resource):
@@ -60,9 +74,13 @@ class TweetsResource(Resource):
     @api.expect(json_new_tweet, validate=True)
     @api.response(422, 'Invalid tweet')
     def post(self):
+        user = load_user_from_request(request)
+        if user is None:
+            return api.abort(401, "You need to provide a valid api_key in the header or request url")
         text = api.payload["text"]
         if len(text) > 0:
             tweet = Tweet(text=text)
+            tweet.user_id = user.id
             db.session.add(tweet)
             db.session.commit()
             return tweet, 201
@@ -73,3 +91,15 @@ class TweetsResource(Resource):
     def get(self):
         tweets = db.session.query(Tweet).all()
         return tweets
+
+
+@login_manager.request_loader
+def load_user_from_request(request):
+    api_key = request.headers.get('api_key')
+    if not api_key:
+        api_key = request.args.get('api_key')
+    if api_key:
+        user = db.session.query(User).filter_by(api_key=api_key).first()
+        if user:
+            return user
+    return None
